Remove deprecated copy-related remnants, move all bounds-related functions to bounds.py.
