+++ flamedisx/nest/lxe_blocks/detection.py
@@ -90,8 +90,9 @@ class DetectPhotonsOrElectrons(fd.Block):
             rvs = [out_bound * np.ones_like(support)
                    for out_bound, support in zip(out_bounds, supports)]
 
-            self.bayes_bounds_binomial(d, self.quanta_name + 's_produced', supports=supports,
-                                       rvs_binom=rvs, ns_binom=ns, ps_binom=ps, bound=bound)
+            fd.bounds.bayes_bounds_binomial(df=d, in_dim=self.quanta_name + 's_produced', supports=supports,
+                                            rvs_binom=rvs, ns_binom=ns, ps_binom=ps,
+                                            bound=bound, bounds_prob=self.source.bounds_prob)
 
 
 @export
