+++ flamedisx/nest/lxe_blocks/secondary_quanta_generation.py
@@ -55,5 +55,6 @@ class MakeS2Photons(fd.Block):
             rvs = [out_bound * np.ones_like(support)
                    for out_bound, support in zip(out_bounds, supports)]
 
-            self.bayes_bounds_normal(d, 'electrons_detected', supports=supports,
-                                     rvs_normal=rvs, mus_normal=mus, sigmas_normal=sigmas, bound=bound)
+            fd.bounds.bayes_bounds_normal(df=d, in_dim='electrons_detected', supports=supports,
+                                          rvs_normal=rvs, mus_normal=mus, sigmas_normal=sigmas,
+                                          bound=bound, bounds_prob=self.source.bounds_prob)
